feat: resolve build cli
