+++ lerna.json
@@ -1,5 +1,5 @@
 {
-  "packages": ["packages/*"],
+  "packages": ["packages/*", "scrpits/*"],
   "npmClient": "yarn",
   "useWorkspaces": true,
   "version": "independent",
