+++ package.json
@@ -4,35 +4,29 @@
   "private": true,
   "workspaces": {
     "packages": [
-      "packages/!(**.bak)*"
+      "packages/!(**.bak)*",
+      "scripts/cli"
     ]
   },
   "scripts": {
     "lint": "eslint --ext .js,.ts,.tsx,.vue ./packages ./site ",
     "lint:style": "stylelint --fix ./packages/**/*.{css,less} --custom-syntax",
-    "lint:fix": "eslint --fix ./packages ./site ./scripts --ext .js,.ts,.tsx,.vue",
+    "lint:fix": "eslint --fix ./packages ./site --ext .js,.ts,.tsx,.vue",
     "test:unit": "jest --passWithNoTests",
     "dev": "vite serve site",
-    "build": "yarn build:library && yarn build:component",
+    "build": "yarn build:dist && yarn build:lib",
     "build:site": "vite build site --mode development",
-    "build:library": "rimraf dist && cross-env NODE_ENV=production ts-node --project=scripts/tsconfig.scripts.json scripts/build-library.ts",
-    "build:component": "rimraf lib && npm run generate:declation && cross-env NODE_ENV=production ts-node --project=scripts/tsconfig.scripts.json scripts/build-component/index.ts",
-    "compile:dev": "node ./node_modules/.bin/ts-node --project=./scripts/tsconfig.scripts.json ./scripts/build-component/index.ts",
+    "build:dist": "rimraf dist && lerna run --scope cli --stream dist",
+    "build:lib": "rimraf lib && lerna run --scope cli --stream lib",
+    "build:release": "rimraf lib && lerna run --scope cli release",
     "cz": "git-cz",
     "generate:icon": "lerna run --scope @bkui-vue/icon build",
-    "release": "yarn build && rimraf package.json.bak && cross-env NODE_ENV=production ts-node --project=scripts/tsconfig.scripts.json scripts/release.ts",
+    "release": "yarn build && rimraf package.json.bak && yarn build:release",
     "changelog": "conventional-changelog -c .changelog.config.js -p angular -i CHANGELOG.md -s -l",
     "cc": "plop --plopfile=./plopfile.js",
-    "prepare": "husky install",
-    "generate:declation": "rimraf lib && tsc -p ./scripts/tsconfig.declaration.json"
+    "prepare": "husky install"
   },
   "devDependencies": {
-    "@babel/core": "~7.17.9",
-    "@babel/eslint-parser": "~7.17.0",
-    "@babel/plugin-transform-runtime": "~7.17.0",
-    "@babel/plugin-transform-typescript": "~7.16.8",
-    "@babel/preset-env": "~7.16.11",
-    "@babel/preset-typescript": "~7.16.7",
     "@commitlint/cli": "~16.2.3",
     "@commitlint/config-conventional": "~16.2.1",
     "@types/jest": "~27.4.1",
@@ -50,8 +44,6 @@
     "@vue/test-utils": "~2.0.0-rc.12",
     "@vue/vue3-jest": "~27.0.0-alpha.4",
     "babel-jest": "~27.5.1",
-    "babel-loader": "~8.2.4",
-    "chalk": "~4.1.2",
     "clipboard": "~2.0.8",
     "commitizen": "~4.2.4",
     "conventional-changelog-cli": "~2.2.2",
@@ -66,20 +58,15 @@
     "eslint-plugin-promise": "~6.0.0",
     "eslint-plugin-simple-import-sort": "~7.0.0",
     "eslint-plugin-vue": "~8.6.0",
-    "file-loader": "~6.2.0",
     "highlight.js": "~11.5.1",
     "husky": "~7.0.1",
     "jest": "~27.5.1",
     "jest-transform-stub": "~2.0.0",
     "lerna": "~4.0.0",
-    "less": "~4.1.1",
     "lint-staged": "~12.3.7",
     "markdown-it-container": "~3.0.0",
     "normalize-wheel": "~1.0.1",
-    "ora": "~5.4.1",
     "plop": "~3.0.5",
-    "postcss": "~8.4.12",
-    "postcss-less": "~6.0.0",
     "prettier": "~2.6.2",
     "prismjs": "~1.27.0",
     "rimraf": "~3.0.2",
@@ -87,20 +74,17 @@
     "stylelint-order": "~5.0.0",
     "stylelint-scss": "~4.2.0",
     "ts-jest": "~27.1.4",
-    "ts-loader": "~9.2.8",
     "ts-node": "~10.7.0",
     "typescript": "~4.6.3",
-    "url-loader": "~4.1.1",
     "vite": "~2.9.1",
     "vite-plugin-md": "~0.12.4",
     "vue": "^3.2.0",
     "vue-router": "~4.0.14",
-    "vue-types": "~4.1.1",
-    "webpack": "~5.72.0"
+    "vue-types": "~4.1.1"
   },
   "peerDependencies": {
-    "vue": "^3.2.0",
-    "highlight.js": "~11.5.0"
+    "highlight.js": "~11.5.0",
+    "vue": "^3.2.0"
   },
   "main": "dist/bkui-vue.umd.js",
   "module": "dist/bkui-vue.esm.js",
