+++ scripts/cli/.eslintrc.js
@@ -27,14 +27,13 @@
 module.exports = {
   root: true,
   extends: [
-    'plugin:vue/vue3-recommended',
     'eslint-config-tencent',
   ],
   parserOptions: {
     parser: '@typescript-eslint/parser',
     ecmaVersion: 2019,
     sourceType: 'module',
-    poject: './tsconfig.json',
+    project: './tsconfig.json',
   },
   plugins: [
     'codecc',
@@ -43,21 +42,11 @@ module.exports = {
   env: {
     es6: true,
     node: true,
-    jest: true,
-    browser: true,
   },
   overrides: [
     {
       files: ['*.ts', '*.tsx'],
       extends: ['eslint-config-tencent/ts'],
-    },
-    {
-      files: ['*.vue'],
-      parser: require.resolve('vue-eslint-parser'),
-      rules: {
-        indent: 'off',
-        'codecc/license': 'off',
-      },
     },
     {
       files: [
@@ -122,21 +111,5 @@ module.exports = {
         ignoreTemplateLiterals: true,
       },
     ],
-    'vue/multi-word-component-names': 'off',
-    'vue/no-setup-props-destructure': 'off',
-    'vue/require-default-prop': 'off',
-    'vue/script-indent': ['error', 2, {
-      baseIndent: 1,
-    }],
-    'vue/component-tags-order': ['error', {
-      order: ['template', 'script', 'style'],
-    }],
-    'vue/component-name-in-template-casing': ['error', 'kebab-case', {
-      registeredComponentsOnly: false,
-      ignores: [],
-    }],
-    'vue/prefer-separate-static-class': 'error',
-    'vue/prefer-true-attribute-shorthand': 'error',
-    'vue/v-on-function-call': 'error',
   },
 };
