+++ scripts/cli/babel.config.js
