+++ scripts/cli/compiler/webpack-script.ts
@@ -24,13 +24,14 @@
 * IN THE SOFTWARE.
 */
 import path, { basename, extname } from 'path';
-import webpack from 'webpack';
+import webpack, { Stats } from 'webpack';
+import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer';
 
-import { ITaskItem } from '../typings/task';
+import { ILibTaskOption, ITaskItem } from '../typings/task';
 
 import { DIST_URL } from './helpers';
 
-export const webpackBuildScript = async (entryList: ITaskItem[]) => {
+export const webpackBuildScript = async (entryList: ITaskItem[], taskOption: ILibTaskOption) => {
   const entry: webpack.EntryObject = {};
   entryList.forEach(({ url }) => {
     if (url && !url.includes('/styles/')
@@ -82,7 +83,7 @@ export const webpackBuildScript = async (entryList: ITaskItem[]) => {
               options: {
                 transpileOnly: false, // Set to true if you are using fork-ts-checker-webpack-plugin
                 projectReferences: true,
-                configFile: path.resolve(__dirname, '../../tsconfig.component.json'),
+                configFile: path.resolve(__dirname, '../tsconfig.component.json'),
               },
             },
           ],
@@ -111,11 +112,11 @@ export const webpackBuildScript = async (entryList: ITaskItem[]) => {
       'vue-types',
       ({ request }, cb)  => {
         const prefix = '@bkui-vue/';
-        if (request.includes('@babel/')) {
-          return cb(null, request);
+        if (request?.includes('@babel/')) {
+          return cb(undefined, request);
         }
-        if (request.indexOf(prefix) === 0) {
-          return cb(null, request.replace(prefix, '../'));
+        if (request?.indexOf(prefix) === 0) {
+          return cb(undefined, request.replace(prefix, '../'));
         }
         cb();
       },
@@ -124,24 +125,29 @@ export const webpackBuildScript = async (entryList: ITaskItem[]) => {
       extensions: ['.tsx', '.ts', '.js'],
     },
     plugins: [
+      taskOption.analyze ? new BundleAnalyzerPlugin({
+        analyzerMode: 'server',
+        analyzerHost: '127.0.0.1',
+        analyzerPort: 8888,
+      }) : undefined,
       new webpack.ProgressPlugin(),
-    ],
+    ].filter(Boolean),
   });
   return new Promise<void>((resolve, reject) => {
-    compiler.run((err: Error, stats: webpack.Stats) => {
+    compiler.run((err: Error | null | undefined, stats: Stats | undefined) => {
       if (err) {
         console.info(err);
         reject(err.message);
         return;
       }
-      if (stats.hasErrors()) {
+      if (stats?.hasErrors()) {
         stats.compilation.errors.forEach((e) => {
           console.log(e.message);
         });
 
         reject('Build failed');
       }
-      console.log('\n', stats.toString({ colors: true }), '\n');
+      console.log('\n', stats?.toString({ colors: true }), '\n');
       resolve();
     });
   });
