+++ scripts/cli/package.json
@@ -1,0 +1,59 @@
+{
+  "name": "cli",
+  "version": "0.0.0",
+  "description": "",
+  "main": "index.ts",
+  "module": "index.ts",
+  "scripts": {
+    "dist": "ts-node index.ts dist",
+    "lib": "ts-node index.ts lib",
+    "release": "ts-node index.ts release"
+  },
+  "devDependencies": {
+    "@babel/core": "~7.17.9",
+    "@babel/plugin-transform-runtime": "~7.17.0",
+    "@babel/plugin-transform-typescript": "~7.16.8",
+    "@babel/preset-env": "~7.16.11",
+    "@babel/preset-typescript": "~7.16.7",
+    "@types/less": "~3.0.3",
+    "@types/node": "~17.0.23",
+    "@types/postcss-less": "~4.0.2",
+    "@types/webpack-bundle-analyzer": "~4.4.1",
+    "@typescript-eslint/eslint-plugin": "~5.19.0",
+    "@typescript-eslint/parser": "~5.19.0",
+    "@vitejs/plugin-vue": "~2.3.1",
+    "@vitejs/plugin-vue-jsx": "~1.3.9",
+    "@vue/babel-plugin-jsx": "~1.1.1",
+    "@vue/babel-preset-jsx": "~1.2.4",
+    "babel-loader": "~8.2.4",
+    "chalk": "~4.1.2",
+    "commander": "~9.2.0",
+    "eslint": "~8.13.0",
+    "eslint-config-tencent": "~1.0.4",
+    "eslint-plugin-codecc": "~0.1.0",
+    "eslint-plugin-import": "~2.26.0",
+    "eslint-plugin-node": "~11.1.0",
+    "eslint-plugin-prettier": "~4.0.0",
+    "eslint-plugin-promise": "~6.0.0",
+    "eslint-plugin-simple-import-sort": "~7.0.0",
+    "eslint-plugin-vue": "~8.6.0",
+    "file-loader": "~6.2.0",
+    "less": "~4.1.1",
+    "ora": "~5.4.1",
+    "postcss": "~8.4.12",
+    "postcss-less": "~6.0.0",
+    "rimraf": "~3.0.2",
+    "ts-loader": "~9.2.8",
+    "ts-node": "~10.7.0",
+    "typescript": "~4.6.3",
+    "url-loader": "~4.1.1",
+    "vite": "~2.9.1",
+    "vue": "^3.2.0",
+    "vue-router": "~4.0.14",
+    "vue-types": "~4.1.1",
+    "webpack": "~5.72.0",
+    "webpack-bundle-analyzer": "~4.5.0"
+  },
+  "author": "",
+  "license": "ISC"
+}
