+++ scripts/cli/tsconfig.component.json
@@ -4,23 +4,20 @@
     "strict": true,
     "moduleResolution": "node",
     "esModuleInterop": true,
-    "declaration": true,
-    "emitDeclarationOnly": true,
-    "declarationDir": "../lib",
     "experimentalDecorators": true,
     "jsx": "preserve",
     "noUnusedParameters": true,
     "noUnusedLocals": true,
     "noImplicitAny": false,
     "module": "esnext",
-    "target": "esnext",
+    "target": "es6",
     "lib": [
       "dom",
       "es2017"
     ],
     "skipLibCheck": true,
     "allowJs": true,
-    "baseUrl": "../packages",
+    "baseUrl": "../../packages",
     "paths": {
       "@bkui-vue/*": [
         "*/src"
@@ -29,15 +26,16 @@
     "importHelpers": true,
     "isolatedModules": true,
     "sourceMap": true,
-    "noEmitOnError": true
+    "noEmitOnError": true,
   },
   "include": [
-    "../packages/*/src",
-    "../packages/**/*.d.ts",
-    "../packages/bkui-vue/index.ts",
-    "../index.d.ts"
+    "../../packages/*/src",
+    "../../packages/**/*.d.ts",
+    "../../packages/bkui-vue/index.ts",
+    "../../index.d.ts"
   ],
   "exclude": [
-    "../packages/icon"
+    "__test__",
+    "node_modules"
   ]
 }
