+++ scripts/cli/tsconfig.declaration.json
@@ -4,20 +4,23 @@
     "strict": true,
     "moduleResolution": "node",
     "esModuleInterop": true,
+    "declaration": true,
+    "emitDeclarationOnly": true,
+    "declarationDir": "../../lib",
     "experimentalDecorators": true,
     "jsx": "preserve",
     "noUnusedParameters": true,
     "noUnusedLocals": true,
     "noImplicitAny": false,
     "module": "esnext",
-    "target": "es6",
+    "target": "esnext",
     "lib": [
       "dom",
       "es2017"
     ],
     "skipLibCheck": true,
     "allowJs": true,
-    "baseUrl": "../packages",
+    "baseUrl": "../../packages",
     "paths": {
       "@bkui-vue/*": [
         "*/src"
@@ -26,16 +29,15 @@
     "importHelpers": true,
     "isolatedModules": true,
     "sourceMap": true,
-    "noEmitOnError": true,
+    "noEmitOnError": true
   },
   "include": [
-    "../packages/*/src",
-    "../packages/**/*.d.ts",
-    "../packages/bkui-vue/index.ts",
-    "../index.d.ts"
+    "../../packages/*/src",
+    "../../packages/**/*.d.ts",
+    "../../packages/bkui-vue/index.ts",
+    "../../index.d.ts"
   ],
   "exclude": [
-    "__test__",
-    "node_modules"
+    "../../packages/icon"
   ]
 }
